flatten generated html
latest versions of chrome freeze when rendering large json objects, where almost the entire time is spent on "Hit Test". From the information I could gather, this is likely due to the multiple layers/nesting of html created, where each is potentially clickable and needs to be hit-tested.

The issue only occurrs on chrome.